requirements.txt
django
djangorestframework
django-import-export
django-multiselectfield
django-recurrence
psycopg2=2.9.9